+++ src/main/app/badge/BadgeCard.stories.tsx
@@ -3,17 +3,49 @@
 import BadgeCard from './BadgeCard.tsx'
 import { Meta, StoryObj } from '@storybook/react-vite'
 import { STORYBOOK_CONTENT } from '../../storybook/storybook-content.ts'
+import { expect, waitFor } from 'storybook/test'
+import CharacterContextProvider from '../character/CharacterContextProvider.tsx'
+import BadgeCount from '../character/BadgeCount.tsx'
+import { getBadgerDb } from '../db/badger-db.ts'
 
 const meta: Meta<typeof BadgeCard> = {
   title: 'badge/BadgeCard',
   component: BadgeCard,
 }
 export default meta
-type StoryType = StoryObj<typeof meta.component>
+type StoryType = StoryObj<typeof BadgeCard>
 
 export const Exploration: StoryType = {
   args: {
     badge: STORYBOOK_CONTENT.getBadge('hangman')
+  },
+}
+
+export const Collect_For_Character: StoryType = {
+  tags: ['interaction'],
+  parameters: { characterKey: 'test1' },
+  args: { badge: STORYBOOK_CONTENT.getBadge('hangman') },
+  render: function Render(args) {
+    const { character } = CharacterContextProvider.useCharacterContext()
+    return <>
+      {character && <BadgeCount character={character}/>}
+      <BadgeCard {...args}/>
+    </>
+  },
+  play: async ({ canvas, userEvent }) => {
+    const complete = await canvas.findByRole('switch')
+    await expect(complete).not.toBeChecked()
+    await expect(canvas.getByText('0 badges')).toBeVisible()
+    await userEvent.click(complete)
+    await waitFor(() => expect(complete).toBeChecked())
+    await expect(await canvas.findByText('1 badge')).toBeVisible()
+    const db = await getBadgerDb()
+    await expect((await db.getCharacter('test1'))?.badges?.hangman?.owned).toBe(true)
+    await userEvent.click(complete)
+    await waitFor(() => expect(complete).not.toBeChecked())
+    await expect(await canvas.findByText('0 badges')).toBeVisible()
+    await expect((await db.getCharacter('test1'))?.badges?.hangman?.owned).toBe(false)
+    await expect((await db.getCharacter('test2'))?.badges).toBeUndefined()
   },
 }