+++ src/main/storybook/SetTheme.tsx
@@ -1,9 +1,11 @@
 import { useColorScheme } from '@mui/joy'
-import { FC, ReactNode } from 'react'
+import { FC, ReactNode, useEffect } from 'react'
 
 const SetTheme: FC<{ children: ReactNode, mode: 'dark' | 'light' }> = ({ children, mode }) => {
   const { setMode } = useColorScheme()
-  setMode(mode)
+  useEffect(() => {
+    setMode(mode)
+  }, [mode, setMode])
   return children
 
 }