+++ .storybook/main.ts
@@ -1,0 +1,19 @@
+// noinspection JSUnusedGlobalSymbols
+
+import type { StorybookConfig } from '@storybook/react-vite'
+
+const config: StorybookConfig = {
+  stories: [
+    '../src/app/**/*.stories.@(ts|tsx|js|jsx)'
+  ],
+  addons: [
+    '@storybook/addon-vitest',
+    '@vueless/storybook-dark-mode'
+  ],
+  framework: {
+    'name': '@storybook/react-vite',
+    'options': {}
+  },
+}
+
+export default config